requirements.txt
requests
pyaudio
python-dotenv
webrtcvad
chromadb
groq
flask
openai
playwright
